backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.6
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
soundfile==0.13.1
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
gunicorn
psycopg2-binary
djangorestframework-simplejwt[crypto]
google-cloud-storage
google-cloud-firestore
